Optimize importing and exporting of components, refactor NotFound component for brevity and add 404 route
