--- src/components/Header/index.js
@@ -1,3 +1,0 @@
-import Header from './Header';
-
-export default Header;
