--- src/components/Preloader/index.js
@@ -1,3 +1,0 @@
-import Preloader from './Preloader';
-
-export default Preloader;
