+++ src/components/index.js
@@ -1,0 +1,11 @@
+import Header from './Header/Header';
+import Preloader from './Preloader/Preloader';
+import LikeButton from './User/LikeButton';
+import LoginButton from './User/LoginButton';
+
+export {
+    Header,
+    Preloader,
+    LikeButton,
+    LoginButton,
+};
