+++ src/containers/NotFound/index.js
@@ -1,8 +1,10 @@
-export default function NotFound() {
-  return (
-    <div className="container">
-      <h1>Doh! 404!</h1>
-      <p>These are <em>not</em> the droids you are looking for!</p>
-    </div>
-  );
-}
+import React from 'react';
+
+const NotFound = () => (
+  <div className="container">
+    <h1>Doh! 404!</h1>
+    <p>These are <em>not</em> the droids you are looking for!</p>
+  </div>
+);
+
+export default NotFound;
