+++ src/containers/index.js
@@ -1,0 +1,9 @@
+import Root from './App/Root';
+import App from './App/App';
+import NotFound from './NotFound';
+
+export {
+    Root,
+    App,
+    NotFound,
+};
