--- src/routes/contact/index.js
@@ -1,3 +1,0 @@
-import contact from './contact';
-
-export default contact;
