--- src/routes/home/index.js
@@ -1,3 +1,0 @@
-import home from './home';
-
-export default home;
